llm-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
instructor==1.0.0
openai==1.10.0
anthropic==0.39.0
supabase>=2.25.1
python-jose[cryptography]==3.3.0
dateparser==1.2.0
websockets>=15.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.27.2,<0.28.0
